pyproject.toml
[project]
name = "agent1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "pydantic-ai-slim[duckduckgo]>=0.1.3",
    "pydantic-ai>=0.1.3",
]
